requirements.txt
fastapi
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic
uvicorn
python-dotenv